Add simple cache module.
It contains a variant of getUnshortened with an additional argument
containing the cache of previously seen urls.

Having written the above, I realized that we also need the ability to
update the cache. :) One way would be to return the updated cache, but I
think it isn't so nice because then we would restrict ourselves too
much (no choice whether to cache the url, multiple threads would require
cache merging). So we could either output a tuple of the result and
maybe a new cache value or return just the result and let someone else
handle the cache.

Well, thinking about *that*, I don't think I should reimplement
memoization, at least not if this weren't such a nice little exercise.
But at least for now I'll ignore it and check later if there's a nice
solution to this problem.